+++ src/progDinamica.h
@@ -1,5 +1,5 @@
 #include "matriz.h" 
 
 int somaMinima(Celula **matriz, int numLinhas, int numColunas);
-int qtdCaminhos(Celula **matriz, int numLinhas, int numColunas);
+unsigned long long int qtdCaminhos(Celula **matriz, int numLinhas, int numColunas);
 bool verificaPosicao(Celula **matriz, int numLinhas, int numColunas,int x, int y);
